# backend/clients/ts-sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Patron API",
    "description": "An open source Patreon alternative with lower fees designed for creators who publish ongoing sequential content like books, podcasts, and comics.",
    "contact": {
      "name": "Patron Team",
      "email": "support@patron.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.patron.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/api-keys": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "List API keys with cursor-based pagination and optional filtering",
        "description": "# Errors\nReturns error if API keys database query fails or connection issues occur",
        "operationId": "list_api_keys",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "Starting point UUID for paginated API key results",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of API keys to return per page (default: 50, max: 100)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "description": "Only show active or inactive API keys",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API keys list retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeysListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to list API keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during API keys listing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Create a new API key",
        "description": "# Errors\nReturns error if permissions are invalid, database error, or key generation fails",
        "operationId": "create_api_key",
        "requestBody": {
          "description": "API key creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid API key data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to create API keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during API key creation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/api-keys/{api_key_id}": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "Get a specific API key by ID",
        "description": "# Errors\nReturns error if API key not found, access denied, or database connection error",
        "operationId": "get_api_key",
        "parameters": [
          {
            "name": "api_key_id",
            "in": "path",
            "description": "UUID of the API key to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to view API keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - API key belongs to different user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "API key not found or access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during API key retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "API Keys"
        ],
        "summary": "Update an API key",
        "description": "# Errors\nReturns error if API key not found, access denied, or database update error",
        "operationId": "update_api_key",
        "parameters": [
          {
            "name": "api_key_id",
            "in": "path",
            "description": "UUID of the API key to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Updated API key data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "API key updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to update API keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - cannot update API key owned by different user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "API key not found for update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during API key update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "API Keys"
        ],
        "summary": "Delete an API key (hard delete for security)",
        "description": "# Errors\nReturns error if API key not found, access denied, or database deletion error",
        "operationId": "delete_api_key",
        "parameters": [
          {
            "name": "api_key_id",
            "in": "path",
            "description": "UUID of the API key to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "API key deleted successfully"
          },
          "401": {
            "description": "Authentication required to delete API keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - cannot delete API key owned by different user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "API key not found for deletion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during API key deletion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/auth/check-email": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Check if email exists",
        "description": "# Errors\nReturns 404 if email doesn't exist, 204 if it does exist, or 500 for database errors.",
        "operationId": "check_email",
        "requestBody": {
          "description": "Email address to check for registration",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Email exists in the system"
          },
          "404": {
            "description": "Email not found in the system"
          },
          "500": {
            "description": "Database connection failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "DATABASE_ERROR",
                  "error": "Database connection failed"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forgot password",
        "description": "# Errors\nReturns an error if database operations fail or email service fails.",
        "operationId": "forgot_password",
        "requestBody": {
          "description": "Email address for password reset request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password reset email sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponse"
                }
              }
            }
          },
          "500": {
            "description": "Password reset service error or email delivery failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "EMAIL_SERVICE_ERROR",
                  "error": "Email service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/google": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Google `OAuth` redirect",
        "description": "# Errors\nReturns an error if session operations fail or `OAuth` service configuration is invalid.",
        "operationId": "google_auth_redirect",
        "responses": {
          "302": {
            "description": "Redirect to Google OAuth consent screen"
          },
          "500": {
            "description": "OAuth initialization failed or session storage error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "OAUTH_SESSION_ERROR",
                  "error": "Failed to initialize OAuth session"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/google/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Google `OAuth` callback",
        "description": "# Errors\nReturns an error if `OAuth` state verification fails, token exchange fails, or database operations fail.",
        "operationId": "google_auth_callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "Authorization code from Google",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "State parameter for CSRF protection",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend application"
          },
          "400": {
            "description": "Invalid authorization code or state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "OAUTH_INVALID_CODE",
                  "error": "Invalid authorization code"
                }
              }
            }
          },
          "500": {
            "description": "OAuth service error or database connection failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "OAUTH_PROVIDER_ERROR",
                  "error": "OAuth provider connection failed"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User login",
        "description": "# Errors\nReturns an error if credentials are invalid, email is not verified, or database operations fail.",
        "operationId": "login",
        "requestBody": {
          "description": "User login credentials including email and password",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials or email not verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_INVALID_CREDENTIALS",
                  "error": "Invalid email or password"
                }
              }
            }
          },
          "500": {
            "description": "Login service error or database connection failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_SERVICE_ERROR",
                  "error": "Authentication service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout",
        "description": "# Errors\nReturns an error if session operations fail.",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Successfully logged out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/api/auth/verify-email": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Email verification",
        "description": "# Errors\nReturns an error if token is invalid, expired, or database operations fail.",
        "operationId": "verify_email",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Email verification token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend after verification"
          },
          "400": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_INVALID_VERIFICATION_TOKEN",
                  "error": "Invalid or expired verification token"
                }
              }
            }
          },
          "500": {
            "description": "Email verification service error or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "VERIFICATION_SERVICE_ERROR",
                  "error": "Email verification service unavailable"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeysListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ApiKeyResponse"
        },
        "description": "List of API keys for a user",
        "example": [
          {
            "createdAt": "2023-01-01T00:00:00Z",
            "expiresAt": "2024-01-01T00:00:00Z",
            "id": "d290f1ee-6c54-4b01-90e6-d701748f0851",
            "isActive": true,
            "keyPrefix": "pk_live_",
            "lastUsedAt": "2023-01-01T12:00:00Z",
            "name": "Production API Key",
            "permissions": [
              "read:posts",
              "write:posts"
            ],
            "updatedAt": "2023-01-01T12:00:00Z",
            "userId": "b2c3d4e5-6f78-9012-bcde-f12345678901"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API key authentication using Bearer token in Authorization header"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid",
        "description": "Session-based authentication using secure HTTP-only cookies"
      }
    }
  },
...
}